+++ lib/src/entity/event/event_param.dart
@@ -7,4 +7,10 @@ class EventParameter {
   factory EventParameter.string(String value) => EventParameter._(value);
   factory EventParameter.number(num value) => EventParameter._(value);
   factory EventParameter.boolean(bool value) => EventParameter._(value);
+  factory EventParameter.dynamic(dynamic value) {
+    if (value is String || value is num || value is bool) {
+      return EventParameter._(value);
+    }
+    throw Exception('Invalid value for EventParameter, `$value`');
+  }
 }